requirements.txt

langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
langgraph==0.2.18
langsmith==0.1.114
pdfplumber
python-dotenv
urllib3
